add styles to shadow roots, too
Where I want to use this extension, only very few
extensions are actually allowed. Shadow root support
for stylus can be added with a custom JS, but it's
unlikely that an extension like Tampermonkey would
be allowed. Therefore this change adds the shadow
root support script as part of Stylus-without-network.